+++ packages/charts/react-charts/library/bundle-size/DeclarativeChart.fixture.js
@@ -1,0 +1,7 @@
+import { DeclarativeChart } from '@fluentui/react-charts';
+
+console.log(DeclarativeChart);
+
+export default {
+  name: 'DeclarativeChart',
+};