+++ packages/react-components/react-toast/library/src/components/ToastContainer/ToastContainer.tsx
@@ -5,6 +5,7 @@ import { useToastContainerStyles_unstable } from './useToastContainerStyles.styl
 import type { ToastContainerProps } from './ToastContainer.types';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
 import { useToastContainerContextValues_unstable } from './useToastContainerContextValues';
+import { useCustomStyleHook_unstable } from '@fluentui/react-shared-contexts';
 
 /**
  * ToastContainer component
@@ -13,6 +14,8 @@ export const ToastContainer: ForwardRefComponent<ToastContainerProps> = React.fo
   const state = useToastContainer_unstable(props, ref);
 
   useToastContainerStyles_unstable(state);
+  useCustomStyleHook_unstable('useToastContainerStyles_unstable')(state);
+
   return renderToastContainer_unstable(state, useToastContainerContextValues_unstable(state));
 });