+++ packages/react-components/react-icons-compat/library/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log - @fluentui/react-icons-compat
 
-This log was last generated on Wed, 19 Mar 2025 15:40:43 GMT and should not be manually modified.
+This log was last generated on Thu, 27 Mar 2025 21:12:51 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.1.11](https://github.com/microsoft/fluentui/tree/@fluentui/react-icons-compat_v0.1.11)
+
+Thu, 27 Mar 2025 21:12:51 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-icons-compat_v0.1.10..@fluentui/react-icons-compat_v0.1.11)
+
+### Patches
+
+- Bump @fluentui/react-jsx-runtime to v9.0.53 ([PR #34034](https://github.com/microsoft/fluentui/pull/34034) by beachball)
+- Bump @fluentui/react-shared-contexts to v9.23.1 ([PR #34034](https://github.com/microsoft/fluentui/pull/34034) by beachball)
+- Bump @fluentui/react-utilities to v9.18.23 ([PR #34034](https://github.com/microsoft/fluentui/pull/34034) by beachball)
 
 ## [0.1.10](https://github.com/microsoft/fluentui/tree/@fluentui/react-icons-compat_v0.1.10)